Use the FactorizedTable's InMemOverflowBuffer in MinMaxAggregateState (#5247)
* Use the FactorizedTable's InMemOverflowBuffer in MinMaxAggregateState

Instead of allocating its own, which can't be freed very easily. Should also reduce data fragmentation, but limits our ability to re-use the space,
though we weren't clearing the overflowBuffer before adding new strings as we could have before.

The MinMaxAggregateState now also tracks the actual size initially
allocated so that it won't allocate more memory if the value is first
replaced by a small, then a slightly larger value
(e.g. for MAX, if "aaaa" -> "b" -> "cc" we can avoid allocating more
memory)

* Remove currentBlock pointer from InMemOverflowBuffer

I think this has been left pointing to the old block when we move the data out of the InMemOverflowBuffer when clearing the AggregateHashTable