Fix malloc-based BM bug causing reads after memory has been freed (#5547)
It was previously possible for an optimistic read to still be reading
when the page is evicted and freed.
It was already unlikely since new reads would clear the marked state and
delay eviction, but was still possible if the page being read is marked
and then evicted in the second pass before the read finishes.
Malloc also usually doesn't immediately return memory to
the OS so the read after free would rarely be a fatal error.
Now a counter tracks the number of ongoing optimistic reads for the page
and we skip evicting if the counter is nonzero.
Once the page state is locked, new reads are blocked, so there is no
data race when evicting.